move files to lib directory
