+++ api/spanish-tips/index.ts
@@ -3,10 +3,10 @@ import {
   addCommentToNotionPage,
   convertToBlockObjectRequest,
   createNotionPage,
-} from "../shared/notion.js";
+} from "../../lib/shared/notion.js";
 import type { Block } from "@tryfabric/martian/build/src/notion/blocks.js";
-import Tip from "./schema.js";
-import { generateData } from "../shared/openai.js";
+import Tip from "../../lib/spanish-tips/schema.js";
+import { generateData } from "../../lib/shared/openai.js";
 import { markdownToBlocks } from "@tryfabric/martian";
 import { waitUntil } from "@vercel/functions";
 import { zodTextFormat } from "openai/helpers/zod";
