+++ lib/shared/openai.ts
