+++ lib/spanish-tips/schema.ts
